Renamed searching to searching-raw
